tests restructured
